THOR-1018 Split the ElasticsearchTransformer into the ES Query Converter and ES Results Converter based on CR feedback.
